Feat/sign

+++ WalletConnectSharp.Core/Models/Relay/PublishOptions.cs
@@ -5,12 +5,12 @@ namespace WalletConnectSharp.Core.Models.Relay
     public class PublishOptions : ProtocolOptionHolder
     {
         [JsonProperty("ttl")]
-        public int TTL { get; set; }
+        public long TTL { get; set; }
         
         [JsonProperty("prompt")]
         public bool Prompt { get; set; }
         
         [JsonProperty("tag")]
-        public int Tag { get; set; }
+        public long Tag { get; set; }
     }
 }
